U1: refactor code into small modules
